internal/api: Remove redundant verification of certificate
This verification only checks whether the fields of a ssh.Certificate
struct have been set to the correct values in the function call before.
This cannot fail other than the compiler making a weird mistake.

cert.SignCert() does not modify fields other than Nonce, SignatureKey
and Signature, therefore nothing can change the certificate and the
verification is a redundant runtime check for compile-time creation of a
struct.